+++ R/dplyr.R
@@ -15,6 +15,14 @@ dplyr::full_join
 #' @export
 dplyr::filter
 
+#' @importFrom dplyr group_by
+#' @export
+dplyr::group_by
+
+#' @importFrom dplyr ungroup
+#' @export
+dplyr::ungroup
+
 #' @importFrom dplyr mutate
 #' @export
 dplyr::mutate
@@ -35,6 +43,16 @@ mutate_PKNCA <- function(.data, ...) {
   dataname <- getDataName(.data)
   .data[[dataname]] <- dplyr::mutate(.data[[dataname]], ...)
   .data
+}
+group_by_PKNCA <- function(.data, ..., .add = FALSE, .drop = group_by_drop_default(.data)) {
+  dataname <- getDataName(.data)
+  .data[[dataname]] <- dplyr::group_by(.data[[dataname]], ..., .add = FALSE, .drop = group_by_drop_default(.data))
+  .data
+}
+ungroup_PKNCA <- function(x, ...) {
+  dataname <- getDataName(x)
+  x[[dataname]] <- dplyr::ungroup(x[[dataname]], ...)
+  x
 }
 
 #' dplyr joins for PKNCA
@@ -66,6 +84,19 @@ right_join.PKNCAconc <- join_maker_PKNCA(dplyr::right_join)
 #' @export
 full_join.PKNCAconc <- join_maker_PKNCA(dplyr::full_join)
 
+#' @rdname inner_join.PKNCAresults
+#' @export
+inner_join.PKNCAdose <- join_maker_PKNCA(dplyr::inner_join)
+#' @rdname inner_join.PKNCAresults
+#' @export
+left_join.PKNCAdose <- join_maker_PKNCA(dplyr::left_join)
+#' @rdname inner_join.PKNCAresults
+#' @export
+right_join.PKNCAdose <- join_maker_PKNCA(dplyr::right_join)
+#' @rdname inner_join.PKNCAresults
+#' @export
+full_join.PKNCAdose <- join_maker_PKNCA(dplyr::full_join)
+
 #' dplyr filtering for PKNCA
 #'
 #' @inheritParams dplyr::filter
@@ -75,6 +106,9 @@ filter.PKNCAresults <- filter_PKNCA
 #' @rdname filter.PKNCAresults
 #' @export
 filter.PKNCAconc <- filter_PKNCA
+#' @rdname filter.PKNCAresults
+#' @export
+filter.PKNCAdose <- filter_PKNCA
 
 #' dplyr mutate-based modification for PKNCA
 #' 
@@ -85,3 +119,29 @@ mutate.PKNCAresults <- mutate_PKNCA
 #' @rdname mutate.PKNCAresults
 #' @export
 mutate.PKNCAconc <- mutate_PKNCA
+#' @rdname mutate.PKNCAresults
+#' @export
+mutate.PKNCAdose <- mutate_PKNCA
+
+#' dplyr grouping for PKNCA
+#' 
+#' @inheritParams dplyr::group_by
+#' @family dplyr verbs
+#' @export
+group_by.PKNCAresults <- group_by_PKNCA
+#' @rdname group_by.PKNCAresults
+#' @export
+group_by.PKNCAconc <- group_by_PKNCA
+#' @rdname group_by.PKNCAresults
+#' @export
+group_by.PKNCAdose <- group_by_PKNCA
+#' @rdname group_by.PKNCAresults
+#' @inheritParams dplyr::ungroup
+#' @export
+ungroup.PKNCAresults <- ungroup_PKNCA
+#' @rdname group_by.PKNCAresults
+#' @export
+ungroup.PKNCAconc <- ungroup_PKNCA
+#' @rdname group_by.PKNCAresults
+#' @export
+ungroup.PKNCAdose <- ungroup_PKNCA