+++ NEWS.md
@@ -32,9 +32,9 @@ the dosing including dose amount and route.
   `conc_above` as a column to the intervals to allow calculation of
   `time_above`.  With this change, the "conc_above" `PKNCA.options()` value has
   been removed.
-* Added dplyr joins, filter, and mutate to allow modification of PKNCA objects
-  after creation.  (Note that these functions will make the provenance no longer
-  match for PKNCAresults objects.)
+* Added dplyr joins, filter, mutate, group_by, and ungroup to allow modification
+  of PKNCA objects after creation.  (Note that these functions will make the
+  provenance no longer match for PKNCAresults objects.)
 
 ## Breaking Changes